BNAS shape pruning processor. (#1903)
### Changes

Handles input masks in NAS by inheriting shape pruning class from common
folder. Current "get_input_mask" function doesn't consider
"add_dynamic_inputs" in elastic width, which can be used to tweak the
elasticity configuration. Currently, in some cases
_get_next_node_sparse_multiplier raises error: "doesn't have closing
mask". This PR fixes this issue.

### Reason for changes

Enable more flexibility in NAS to tweak elastic configuration. 

### Related tickets

N/A

### Tests

Included tests for dynamic inputs.

---------

+++ nncf/experimental/torch/nas/bootstrapNAS/search/search.py
@@ -287,7 +287,7 @@ class SearchAlgorithm(BaseSearchAlgorithm):
         return self._vars_upper
 
     @property
-    def num_vars(self) -> float:
+    def num_vars(self) -> int:
         """
         Number of design variables used by the search algorithm.
         :return:
